+++ climpred/tests/test_graphics.py
@@ -1,25 +1,47 @@
+import numpy as np
+
 from climpred.bootstrap import bootstrap_perfect_model
 from climpred.graphics import plot_bootstrapped_skill_over_leadyear
-from climpred.tutorial import load_dataset
+
+BOOTSTRAP = 3
 
 
-def test_mpi_pm_plot_bootstrapped_skill_over_leadyear():
+def test_mpi_pm_plot_bootstrapped_skill_over_leadyear_da(
+    PM_da_initialized_1d, PM_da_control_1d
+):
     """
-    Checks plots from bootstrap_perfect_model works.
+    Checks plots from bootstrap_perfect_model works for xr.DataArray.
     """
-    da = load_dataset("MPI-PM-DP-1D").isel(area=1, period=-1)
-    PM_da_ds1d = da["tos"]
-
-    da = load_dataset("MPI-control-1D").isel(area=1, period=-1)
-    PM_da_control1d = da["tos"]
-
-    bootstrap = 3
     res = bootstrap_perfect_model(
-        PM_da_ds1d,
-        PM_da_control1d,
-        metric="pearson_r",
-        dim="init",
-        bootstrap=bootstrap,
-    ).mean("member")
-    res_ax = plot_bootstrapped_skill_over_leadyear(res, 95)
+        PM_da_initialized_1d, PM_da_control_1d, metric='pearson_r', bootstrap=BOOTSTRAP,
+    )
+    res_ax = plot_bootstrapped_skill_over_leadyear(res)
+    assert res_ax is not None
+
+
+def test_mpi_pm_plot_bootstrapped_skill_over_leadyear_single_uninit_lead(
+    PM_da_initialized_1d, PM_da_control_1d
+):
+    """
+    Checks plots from bootstrap_perfect_model works for xr.DataArray.
+    """
+    res = bootstrap_perfect_model(
+        PM_da_initialized_1d, PM_da_control_1d, metric='pearson_r', bootstrap=BOOTSTRAP,
+    )
+    # set all but first uninit lead to nan
+    res[:, 2, 1:] = [np.nan] * (res.lead.size - 1)
+    res_ax = plot_bootstrapped_skill_over_leadyear(res)
+    assert res_ax is not None
+
+
+def test_mpi_pm_plot_bootstrapped_skill_over_leadyear_ds(
+    PM_ds_initialized_1d, PM_ds_control_1d
+):
+    """
+    Checks plots from bootstrap_perfect_model works for xr.Dataset with one variable.
+    """
+    res = bootstrap_perfect_model(
+        PM_ds_initialized_1d, PM_ds_control_1d, metric='pearson_r', bootstrap=BOOTSTRAP,
+    )
+    res_ax = plot_bootstrapped_skill_over_leadyear(res)
     assert res_ax is not None